Split AbilityRoller into ManualEntry + DiceRoller sub-components
AbilityRoller was 866 lines with two independent UI modes sharing no
state. Extract ManualEntry (~180 lines) and DiceRoller (~340 lines),
leaving AbilityRoller as a thin ~30-line wrapper that toggles between them.
